feat!: freeze Bus and Scheduler registration signatures (#1301)
Locks down Bus and Scheduler registration signatures ahead of the v1.0
API freeze. Full rationale and edge-case analysis in
`design/specs/007-keyword-only-api-freeze/design.md`.

### Notable Changes
- `on_error` now passes through on 10 previously-incomplete Bus delegate
methods (`on_homeassistant_*`, `on_websocket_*`, `on_app_*`,
`on_hassette_service_failed/crashed/started`), so per-listener error
handling no longer requires dropping down to the underlying primary
method
- `App.cleanup`, `Resource.cleanup`, and `DatabaseService.cleanup` now
type `timeout` as `float | None` instead of `int | None`, matching every
other timeout in the codebase

### Keyword-only, required `name`
- All Bus and Scheduler registration methods (and their sync facades)
now require `name` as a keyword-only argument with no default — Pyright
reports an error at development time when it's omitted, closing the gap
where the type said `str | None = None` but the runtime enforced it as
required anyway
- Scheduler methods (`schedule`, `run_in`, `run_once`, `run_every`,
`run_minutely`, `run_hourly`, `run_daily`, `run_cron`) move `name`,
`group`, `jitter`, `timeout`, `timeout_disabled` behind `*`; `run_once`
also makes `if_past` keyword-only
- The three `on_homeassistant_*` Bus methods gain a `*` separator they
previously lacked
- Added `SchedulerNameRequiredError`, mirroring the existing
`ListenerNameRequiredError` — raised at runtime when a dynamic caller
passes an empty string instead of a real name
- `Scheduler.add_job()`/`_add_job()` now validates `job.name` is
non-empty at entry, matching `Bus.add_listener()`'s existing guard

### Remove `name_auto` infrastructure
- Scheduler no longer auto-derives a job name from the callable and
trigger when none is provided — this diverged from the Bus philosophy
and had to be re-explained every time; removed `name_auto` from
`ScheduledJob`, `ScheduledJobRegistration`, and `JobSummary`
- Migration `010.sql` drops the `name_auto` column from
`scheduled_jobs`; existing rows keep their derived name string, only the
metadata flag disappears
- Removed the `nameAutoHint` prop and its render block from the frontend
(`handler-detail-layout.tsx`, `job-detail.tsx`)

### Regression prevention
- New pre-commit linter (`tools/check_registration_signatures.py`)
statically verifies that every public `Bus`/`Scheduler` method with a
`name` parameter keeps it keyword-only with no default, so a future edit
can't silently reopen the positional-arg or optional-name gap this PR
closes
- Updated all docs snippets, migration guide, and internal call sites
(e.g. `StateProxy.on_initialize`) to pass `name=` explicitly

### Housekeeping
- Deduped callable-name lookup and test boilerplate flagged by the
clean-code check
- Fixed stale name-requirement prose in docs after the freeze landed

BREAKING CHANGE: `name` is now a required keyword-only argument on every
Bus and
Scheduler registration method (`on_state_change`, `on_attribute_change`,
`on_call_service`, `on`, `on_homeassistant_*`, `on_websocket_*`,
`on_app_*`,
`on_hassette_service_*`, `schedule`, `run_in`, `run_once`, `run_every`,
`run_minutely`, `run_hourly`, `run_daily`, `run_cron`) and their sync
facades.
Omitting `name` now raises a `TypeError` at call time instead of the old
`ListenerNameRequiredError` — Pyright will catch this at development
time for
static callers. Passing an empty string for `name` still raises
`ListenerNameRequiredError` (Bus) or the new
`SchedulerNameRequiredError`
(Scheduler). Positional arguments after the leading required parameters
(`handler`/`func`, `where`, `trigger`, etc.) are now rejected — all
optional
parameters must be passed by keyword. Update call sites to pass
`name="..."`
explicitly and convert any positional `group=`/`jitter=`/`timeout=`/
`timeout_disabled=`/`if_past=` arguments to keyword form.

Closes #1228
Closes #973
Closes #1238

+++ src/hassette/app/app.py
@@ -149,7 +149,7 @@ class App(Generic[AppConfigT], Resource, metaclass=FinalMeta):
         return date_utils.now()
 
     @final
-    async def cleanup(self, timeout: int | None = None) -> None:
+    async def cleanup(self, timeout: float | None = None) -> None:
         """Cleanup resources owned by the instance.
 
         This method is called during shutdown to cancel tasks and close caches.
